Abstract out test running code
