+++ test/RequesterT.hs
@@ -1,4 +1,4 @@
-{-# LANGUAGE FlexibleContexts, GADTs #-}
+{-# LANGUAGE FlexibleContexts, GADTs, RankNTypes, ScopedTypeVariables #-}
 module Main where
 
 import Control.Monad
@@ -15,6 +15,7 @@ import Reflex.Host.Class
 import Reflex.PerformEvent.Base
 import Reflex.Requester.Base
 import Reflex.Requester.Class
+import Test.Run
 
 import System.Exit (exitFailure, exitSuccess)
 
@@ -23,39 +24,23 @@ data RequestInt a where
 
 main :: IO ()
 main = do
-  (Just out1, Nothing) <- runApp app1
+  (Just out1, Nothing) <- runApp $ unwrapApp app1
   print out1
-  (Just out2, Just (out3, out4)) <- runApp app2
+  (Just out2, Just (out3, out4)) <- runApp $ unwrapApp app2
   print out2
   print out3
   print out4
   when (sort out1 /= out1 || map (+1) out2 /= out3 || out3 /= out4) exitFailure
 
-runApp :: (t ~ SpiderTimeline Global, m ~ SpiderHost Global)
-       => (Event t () -> RequesterT t RequestInt Identity (PerformEventT t m) (Maybe (Ref m (Maybe (EventTrigger t ())))))
-       -> IO (Maybe [Int], Maybe ([Int], [Int]))
-runApp x = runSpiderHost $ do
-  (pulseE, pulseTriggerRef) <- newEventWithTriggerRef
-  ((mactuateRef, e), FireCommand fire) <- hostPerformEventT $ runRequesterT (x pulseE) never
-  hnd <- subscribeEvent (map unwrapRequest . DMap.toList <$> e)
-  out1 <- fireEventRefAndRead pulseTriggerRef () hnd
-  out2 <- fmap join $ forM mactuateRef $ \actuateRef -> do
-    let readPhase = do
-          mGetValue <- readEvent hnd
-          case mGetValue of
-            Nothing -> return Nothing
-            Just getValue -> fmap Just getValue
-    mactuate <- readRef actuateRef
-    mpulse <- readRef pulseTriggerRef
-    forM ((,) <$> mactuate <*> mpulse) $ \(actuate, pulse) -> do
-      _ <- fire [actuate :=> Identity ()] $ return ()
-      [Just a] <- fire [pulse :=> Identity ()] $ readPhase
-      [Just b, Nothing] <- fire [actuate :=> Identity (), pulse :=> Identity ()] $ readPhase
-      return (a,b)
-  return (out1, out2)
-
 unwrapRequest :: DSum tag RequestInt -> Int
 unwrapRequest (_ :=> RequestInt i) = i
+
+unwrapApp :: forall x t a response. (Reflex t, ReflexHost t)
+          => (x -> RequesterT t RequestInt response (PerformEventT t (SpiderHost Global)) a)
+          -> x -> PerformEventT t (SpiderHost Global) (a, Event t [Int])
+unwrapApp x e = do
+  (a, b) <- runRequesterT (x e) never
+  return (a, fmap (map unwrapRequest . DMap.toList) b)
 
 app1 :: (Reflex t, Ref m ~ IORef, Requester t m, Response m ~ Identity, Request m ~ RequestInt) => Event t () -> m (Maybe (Ref m (Maybe (EventTrigger t ()))))
 app1 e = do
